Move css files to module css files
