+++ packages/styleguide/src/screens/Patterns/Carousel/Carousel.module.css
