+++ packages/styleguide/src/screens/Patterns/Figure/Figure.module.css
