+++ packages/styleguide/src/screens/Patterns/FittedImage/FittedImage.module.css
