+++ packages/styleguide/src/screens/Patterns/LeftRight/LeftRight.module.css
