+++ packages/styleguide/src/screens/Patterns/Medallion/Medallion.module.css
